Pull up the fixed_table usage in Lease.{replace,lookup,remove}.
The idea is to make the caller responsible for the fixed_table, making Lease
simpler.

Next step is to clear Lease.addr_in_range so we can remove the range from
Lease.database.

with '#' will be ignored, and an empty message aborts the commit.  # On branch
splitlease # Changes to be committed: # modified: lib/dhcp_server.ml # modified:
lib/lease.ml #